pyproject.toml
[tool.poetry]
name = "daisy-hackathon"
version = "0.1.0"
description = ""
authors = ["bhavanichandra <bhavanichandra9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.1"
boto3 = "^1.24.42"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
